// frontend/src/types/api.ts/genreViewModel.ts
/**
 * Generated by orval v7.10.0 🍺
 * Do not edit manually.
 * Sahib Game Store
 * OpenAPI spec version: v1
 */

export interface GenreViewModel {
  id?: string;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  readonly gamesOfThisGenre?: readonly unknown[] | null;
}